First changes for recording

+++ src/com/prototest/prima/DataStructures/BatteryStats.java
@@ -1,7 +1,7 @@
 package com.prototest.prima.DataStructures;
 
 public class BatteryStats {
-	public long percent;
+	public float percent;
 	public long temp;
 	public long voltage;
 }
